Revert "(refactor) Spread repo_display across 3 files"
